Final Commit

+++ src/components/Country.jsx
@@ -6,7 +6,7 @@ import ListGroup from 'react-bootstrap/ListGroup';
 import { TrashFill, ArrowClockwise, Save } from 'react-bootstrap-icons';
 
 const Country = (props) => {
-  const { country, medals, onIncrement, onDecrement, onDelete, onSave, onReset } = props;
+  const { country, medals, onIncrement, onDecrement, onDelete, onSave, onReset, canDelete, canPatch } = props;
   const getMedalsTotal = (country, medals) => {
     let sum = 0;
     // use medal count displayed in the web page for medal count totals
@@ -43,7 +43,7 @@ const Country = (props) => {
               <ArrowClockwise onClick={ () => onReset(country.id) }></ArrowClockwise>
             </React.Fragment>
             :
-            <TrashFill onClick={() => onDelete(country.id)} className='icon-btn' style={{ color:'red' }} />
+            canDelete && <TrashFill onClick={() => onDelete(country.id)} className='icon-btn' style={{ color:'red' }} />
           }
           </Card.Title>
           <ListGroup variant="flush">
@@ -52,6 +52,7 @@ const Country = (props) => {
             <Medal 
               country={ country } 
               medal={ medal } 
+              canPatch={ canPatch }
               onIncrement={ onIncrement } 
               onDecrement={ onDecrement } />
           </ListGroup.Item>
